Network.AWS.WAFRegional.CreateRule

Description

Creates a Rule , which contains the IPSet objects, ByteMatchSet objects, and other predicates that identify the requests that you want to block. If you add more than one predicate to a Rule , a request must match all of the specifications to be allowed or blocked. For example, suppose you add the following to a Rule :

An IPSet that matches the IP address 192.0.2.44/32

A ByteMatchSet that matches BadBot in the User-Agent header

You then add the Rule to a WebACL and specify that you want to blocks requests that satisfy the Rule . For a request to be blocked, it must come from the IP address 192.0.2.44 and the User-Agent header in the request must contain the value BadBot .

To create and configure a Rule , perform the following steps:

Create and update the predicates that you want to include in the Rule . For more information, see CreateByteMatchSet , CreateIPSet , and CreateSqlInjectionMatchSet .

Use GetChangeToken to get the change token that you provide in the ChangeToken parameter of a CreateRule request.

Submit a CreateRule request.

Use GetChangeToken to get the change token that you provide in the ChangeToken parameter of an UpdateRule request.

Submit an UpdateRule request to specify the predicates that you want to include in the Rule .

Create and update a WebACL that contains the Rule . For more information, see CreateWebACL .

For more information about how to use the AWS WAF API to allow or block HTTP requests, see the AWS WAF Developer Guide .

Creating a Request

createRule

Arguments

:: Text
crName

-> Text
crMetricName

-> Text
crChangeToken

-> CreateRule

Creates a value of CreateRule with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

crName - A friendly name or description of the Rule . You can't change the name of a Rule after you create it.

crMetricName - A friendly name or description for the metrics for this Rule . The name can contain only alphanumeric characters (A-Z, a-z, 0-9); the name can't contain whitespace. You can't change the name of the metric after you create the Rule .

crChangeToken - The value returned by the most recent call to GetChangeToken .

data CreateRule

See: createRule smart constructor.

Request Lenses

crName :: Lens' CreateRule Text

A friendly name or description of the Rule . You can't change the name of a Rule after you create it.

crMetricName :: Lens' CreateRule Text

A friendly name or description for the metrics for this Rule . The name can contain only alphanumeric characters (A-Z, a-z, 0-9); the name can't contain whitespace. You can't change the name of the metric after you create the Rule .

crChangeToken :: Lens' CreateRule Text

The value returned by the most recent call to GetChangeToken .

Destructuring the Response

createRuleResponse

Arguments

:: Int
crrsResponseStatus

-> CreateRuleResponse

Creates a value of CreateRuleResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

crrsRule - The Rule returned in the CreateRule response.

crrsChangeToken - The ChangeToken that you used to submit the CreateRule request. You can also use this value to query the status of the request. For more information, see GetChangeTokenStatus .

crrsResponseStatus - -- | The response status code.

data CreateRuleResponse

See: createRuleResponse smart constructor.

Response Lenses

crrsRule :: Lens' CreateRuleResponse (Maybe Rule)

The Rule returned in the CreateRule response.

crrsChangeToken :: Lens' CreateRuleResponse (Maybe Text)

The ChangeToken that you used to submit the CreateRule request. You can also use this value to query the status of the request. For more information, see GetChangeTokenStatus .

crrsResponseStatus :: Lens' CreateRuleResponse Int

- | The response status code.