Network.AWS.WAFRegional.CreateWebACL

Description

Creates a WebACL , which contains the Rules that identify the CloudFront web requests that you want to allow, block, or count. AWS WAF evaluates Rules in order based on the value of Priority for each Rule .

You also specify a default action, either ALLOW or BLOCK . If a web request doesn't match any of the Rules in a WebACL , AWS WAF responds to the request with the default action.

To create and configure a WebACL , perform the following steps:

Create and update the ByteMatchSet objects and other predicates that you want to include in Rules . For more information, see CreateByteMatchSet , UpdateByteMatchSet , CreateIPSet , UpdateIPSet , CreateSqlInjectionMatchSet , and UpdateSqlInjectionMatchSet .

Create and update the Rules that you want to include in the WebACL . For more information, see CreateRule and UpdateRule .

Use GetChangeToken to get the change token that you provide in the ChangeToken parameter of a CreateWebACL request.

Submit a CreateWebACL request.

Use GetChangeToken to get the change token that you provide in the ChangeToken parameter of an UpdateWebACL request.

Submit an UpdateWebACL request to specify the Rules that you want to include in the WebACL , to specify the default action, and to associate the WebACL with a CloudFront distribution.

For more information about how to use the AWS WAF API, see the AWS WAF Developer Guide .

Creating a Request

createWebACL

Arguments

:: Text
cwaName

-> Text
cwaMetricName

-> WafAction
cwaDefaultAction

-> Text
cwaChangeToken

-> CreateWebACL

Creates a value of CreateWebACL with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cwaName - A friendly name or description of the WebACL . You can't change Name after you create the WebACL .

cwaMetricName - A friendly name or description for the metrics for this WebACL . The name can contain only alphanumeric characters (A-Z, a-z, 0-9); the name can't contain whitespace. You can't change MetricName after you create the WebACL .

cwaDefaultAction - The action that you want AWS WAF to take when a request doesn't match the criteria specified in any of the Rule objects that are associated with the WebACL .

cwaChangeToken - The value returned by the most recent call to GetChangeToken .

data CreateWebACL

See: createWebACL smart constructor.

Request Lenses

cwaName :: Lens' CreateWebACL Text

A friendly name or description of the WebACL . You can't change Name after you create the WebACL .

cwaMetricName :: Lens' CreateWebACL Text

A friendly name or description for the metrics for this WebACL . The name can contain only alphanumeric characters (A-Z, a-z, 0-9); the name can't contain whitespace. You can't change MetricName after you create the WebACL .

cwaDefaultAction :: Lens' CreateWebACL WafAction

The action that you want AWS WAF to take when a request doesn't match the criteria specified in any of the Rule objects that are associated with the WebACL .

cwaChangeToken :: Lens' CreateWebACL Text

The value returned by the most recent call to GetChangeToken .

Destructuring the Response

createWebACLResponse

Arguments

:: Int
cwarsResponseStatus

-> CreateWebACLResponse

Creates a value of CreateWebACLResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cwarsWebACL - The WebACL returned in the CreateWebACL response.

cwarsChangeToken - The ChangeToken that you used to submit the CreateWebACL request. You can also use this value to query the status of the request. For more information, see GetChangeTokenStatus .

cwarsResponseStatus - -- | The response status code.

data CreateWebACLResponse

See: createWebACLResponse smart constructor.

Response Lenses

cwarsWebACL :: Lens' CreateWebACLResponse (Maybe WebACL)

The WebACL returned in the CreateWebACL response.

cwarsChangeToken :: Lens' CreateWebACLResponse (Maybe Text)

The ChangeToken that you used to submit the CreateWebACL request. You can also use this value to query the status of the request. For more information, see GetChangeTokenStatus .

cwarsResponseStatus :: Lens' CreateWebACLResponse Int

- | The response status code.