Network.AWS.WAFRegional.CreateXSSMatchSet

Description

Creates an XssMatchSet , which you use to allow, block, or count requests that contain cross-site scripting attacks in the specified part of web requests. AWS WAF searches for character sequences that are likely to be malicious strings.

To create and configure an XssMatchSet , perform the following steps:

Use GetChangeToken to get the change token that you provide in the ChangeToken parameter of a CreateXssMatchSet request.

Submit a CreateXssMatchSet request.

Use GetChangeToken to get the change token that you provide in the ChangeToken parameter of an UpdateXssMatchSet request.

Submit an UpdateXssMatchSet request to specify the parts of web requests in which you want to allow, block, or count cross-site scripting attacks.

For more information about how to use the AWS WAF API to allow or block HTTP requests, see the AWS WAF Developer Guide .

Creating a Request

createXSSMatchSet

Arguments

:: Text
cxmsName

-> Text
cxmsChangeToken

-> CreateXSSMatchSet

Creates a value of CreateXSSMatchSet with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cxmsName - A friendly name or description for the XssMatchSet that you're creating. You can't change Name after you create the XssMatchSet .

cxmsChangeToken - The value returned by the most recent call to GetChangeToken .

data CreateXSSMatchSet

A request to create an XssMatchSet .

See: createXSSMatchSet smart constructor.

Request Lenses

cxmsName :: Lens' CreateXSSMatchSet Text

A friendly name or description for the XssMatchSet that you're creating. You can't change Name after you create the XssMatchSet .

cxmsChangeToken :: Lens' CreateXSSMatchSet Text

The value returned by the most recent call to GetChangeToken .

Destructuring the Response

createXSSMatchSetResponse

Arguments

:: Int
cxmsrsResponseStatus

-> CreateXSSMatchSetResponse

Creates a value of CreateXSSMatchSetResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cxmsrsXSSMatchSet - An XssMatchSet .

cxmsrsChangeToken - The ChangeToken that you used to submit the CreateXssMatchSet request. You can also use this value to query the status of the request. For more information, see GetChangeTokenStatus .

cxmsrsResponseStatus - -- | The response status code.

data CreateXSSMatchSetResponse

The response to a CreateXssMatchSet request.

See: createXSSMatchSetResponse smart constructor.

Response Lenses

cxmsrsXSSMatchSet :: Lens' CreateXSSMatchSetResponse (Maybe XSSMatchSet)

An XssMatchSet .

cxmsrsChangeToken :: Lens' CreateXSSMatchSetResponse (Maybe Text)

The ChangeToken that you used to submit the CreateXssMatchSet request. You can also use this value to query the status of the request. For more information, see GetChangeTokenStatus .

cxmsrsResponseStatus :: Lens' CreateXSSMatchSetResponse Int

- | The response status code.