Network.AWS.WAFRegional.DeleteSizeConstraintSet

Description

Permanently deletes a SizeConstraintSet . You can't delete a SizeConstraintSet if it's still used in any Rules or if it still includes any SizeConstraint objects (any filters).

If you just want to remove a SizeConstraintSet from a Rule , use UpdateRule .

To permanently delete a SizeConstraintSet , perform the following steps:

Update the SizeConstraintSet to remove filters, if any. For more information, see UpdateSizeConstraintSet .

Use GetChangeToken to get the change token that you provide in the ChangeToken parameter of a DeleteSizeConstraintSet request.

Submit a DeleteSizeConstraintSet request.

Creating a Request

deleteSizeConstraintSet

Arguments

:: Text
dscsSizeConstraintSetId

-> Text
dscsChangeToken

-> DeleteSizeConstraintSet

Creates a value of DeleteSizeConstraintSet with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dscsSizeConstraintSetId - The SizeConstraintSetId of the SizeConstraintSet that you want to delete. SizeConstraintSetId is returned by CreateSizeConstraintSet and by ListSizeConstraintSets .

dscsChangeToken - The value returned by the most recent call to GetChangeToken .

data DeleteSizeConstraintSet

See: deleteSizeConstraintSet smart constructor.

Request Lenses

dscsSizeConstraintSetId :: Lens' DeleteSizeConstraintSet Text

The SizeConstraintSetId of the SizeConstraintSet that you want to delete. SizeConstraintSetId is returned by CreateSizeConstraintSet and by ListSizeConstraintSets .

dscsChangeToken :: Lens' DeleteSizeConstraintSet Text

The value returned by the most recent call to GetChangeToken .

Destructuring the Response

deleteSizeConstraintSetResponse

Arguments

:: Int
dscsrsResponseStatus

-> DeleteSizeConstraintSetResponse

Creates a value of DeleteSizeConstraintSetResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dscsrsChangeToken - The ChangeToken that you used to submit the DeleteSizeConstraintSet request. You can also use this value to query the status of the request. For more information, see GetChangeTokenStatus .

dscsrsResponseStatus - -- | The response status code.

data DeleteSizeConstraintSetResponse

See: deleteSizeConstraintSetResponse smart constructor.

Response Lenses

dscsrsChangeToken :: Lens' DeleteSizeConstraintSetResponse (Maybe Text)

The ChangeToken that you used to submit the DeleteSizeConstraintSet request. You can also use this value to query the status of the request. For more information, see GetChangeTokenStatus .

dscsrsResponseStatus :: Lens' DeleteSizeConstraintSetResponse Int

- | The response status code.