Network.AWS.WAFRegional.DeleteXSSMatchSet

Description

Permanently deletes an XssMatchSet . You can't delete an XssMatchSet if it's still used in any Rules or if it still contains any XssMatchTuple objects.

If you just want to remove an XssMatchSet from a Rule , use UpdateRule .

To permanently delete an XssMatchSet from AWS WAF, perform the following steps:

Update the XssMatchSet to remove filters, if any. For more information, see UpdateXssMatchSet .

Use GetChangeToken to get the change token that you provide in the ChangeToken parameter of a DeleteXssMatchSet request.

Submit a DeleteXssMatchSet request.

Creating a Request

deleteXSSMatchSet

Arguments

:: Text
dxmsXSSMatchSetId

-> Text
dxmsChangeToken

-> DeleteXSSMatchSet

Creates a value of DeleteXSSMatchSet with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dxmsXSSMatchSetId - The XssMatchSetId of the XssMatchSet that you want to delete. XssMatchSetId is returned by CreateXssMatchSet and by ListXssMatchSets .

dxmsChangeToken - The value returned by the most recent call to GetChangeToken .

data DeleteXSSMatchSet

A request to delete an XssMatchSet from AWS WAF.

See: deleteXSSMatchSet smart constructor.

Request Lenses

dxmsXSSMatchSetId :: Lens' DeleteXSSMatchSet Text

The XssMatchSetId of the XssMatchSet that you want to delete. XssMatchSetId is returned by CreateXssMatchSet and by ListXssMatchSets .

dxmsChangeToken :: Lens' DeleteXSSMatchSet Text

The value returned by the most recent call to GetChangeToken .

Destructuring the Response

deleteXSSMatchSetResponse

Arguments

:: Int
dxmsrsResponseStatus

-> DeleteXSSMatchSetResponse

Creates a value of DeleteXSSMatchSetResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dxmsrsChangeToken - The ChangeToken that you used to submit the DeleteXssMatchSet request. You can also use this value to query the status of the request. For more information, see GetChangeTokenStatus .

dxmsrsResponseStatus - -- | The response status code.

data DeleteXSSMatchSetResponse

The response to a request to delete an XssMatchSet from AWS WAF.

See: deleteXSSMatchSetResponse smart constructor.

Response Lenses

dxmsrsChangeToken :: Lens' DeleteXSSMatchSetResponse (Maybe Text)

The ChangeToken that you used to submit the DeleteXssMatchSet request. You can also use this value to query the status of the request. For more information, see GetChangeTokenStatus .

dxmsrsResponseStatus :: Lens' DeleteXSSMatchSetResponse Int

- | The response status code.