Network.AWS.WAFRegional.DisassociateWebACL

Description

Removes a web ACL from the specified resource.

Creating a Request

disassociateWebACL

Arguments

:: Text
dwaResourceARN

-> DisassociateWebACL

Creates a value of DisassociateWebACL with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dwaResourceARN - The ARN (Amazon Resource Name) of the resource from which the web ACL is being removed.

data DisassociateWebACL

See: disassociateWebACL smart constructor.

Request Lenses

dwaResourceARN :: Lens' DisassociateWebACL Text

The ARN (Amazon Resource Name) of the resource from which the web ACL is being removed.

Destructuring the Response

disassociateWebACLResponse

Arguments

:: Int
dwaclrsResponseStatus

-> DisassociateWebACLResponse

Creates a value of DisassociateWebACLResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dwaclrsResponseStatus - -- | The response status code.

data DisassociateWebACLResponse

See: disassociateWebACLResponse smart constructor.

Response Lenses

dwaclrsResponseStatus :: Lens' DisassociateWebACLResponse Int

- | The response status code.