Network.AWS.WAFRegional.GetRateBasedRuleManagedKeys

Description

Returns an array of IP addresses currently being blocked by the RateBasedRule that is specified by the RuleId . The maximum number of managed keys that will be blocked is 10,000. If more than 10,000 addresses exceed the rate limit, the 10,000 addresses with the highest rates will be blocked.

Creating a Request

getRateBasedRuleManagedKeys

Arguments

:: Text
grbrmkRuleId

-> GetRateBasedRuleManagedKeys

Creates a value of GetRateBasedRuleManagedKeys with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

grbrmkNextMarker - A null value and not currently used. Do not include this in your request.

grbrmkRuleId - The RuleId of the RateBasedRule for which you want to get a list of ManagedKeys . RuleId is returned by CreateRateBasedRule and by ListRateBasedRules .

data GetRateBasedRuleManagedKeys

See: getRateBasedRuleManagedKeys smart constructor.

Request Lenses

grbrmkNextMarker :: Lens' GetRateBasedRuleManagedKeys (Maybe Text)

A null value and not currently used. Do not include this in your request.

grbrmkRuleId :: Lens' GetRateBasedRuleManagedKeys Text

The RuleId of the RateBasedRule for which you want to get a list of ManagedKeys . RuleId is returned by CreateRateBasedRule and by ListRateBasedRules .

Destructuring the Response

getRateBasedRuleManagedKeysResponse

Arguments

:: Int
grbrmkrsResponseStatus

-> GetRateBasedRuleManagedKeysResponse

Creates a value of GetRateBasedRuleManagedKeysResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

grbrmkrsNextMarker - A null value and not currently used.

grbrmkrsManagedKeys - An array of IP addresses that currently are blocked by the specified RateBasedRule .

grbrmkrsResponseStatus - -- | The response status code.

data GetRateBasedRuleManagedKeysResponse

See: getRateBasedRuleManagedKeysResponse smart constructor.

Response Lenses

grbrmkrsNextMarker :: Lens' GetRateBasedRuleManagedKeysResponse (Maybe Text)

A null value and not currently used.

grbrmkrsManagedKeys :: Lens' GetRateBasedRuleManagedKeysResponse [Text]

An array of IP addresses that currently are blocked by the specified RateBasedRule .

grbrmkrsResponseStatus :: Lens' GetRateBasedRuleManagedKeysResponse Int

- | The response status code.