Network.AWS.WAFRegional.GetSizeConstraintSet

Description

Returns the SizeConstraintSet specified by SizeConstraintSetId .

Creating a Request

getSizeConstraintSet

Arguments

:: Text
gscsSizeConstraintSetId

-> GetSizeConstraintSet

Creates a value of GetSizeConstraintSet with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gscsSizeConstraintSetId - The SizeConstraintSetId of the SizeConstraintSet that you want to get. SizeConstraintSetId is returned by CreateSizeConstraintSet and by ListSizeConstraintSets .

data GetSizeConstraintSet

See: getSizeConstraintSet smart constructor.

Request Lenses

gscsSizeConstraintSetId :: Lens' GetSizeConstraintSet Text

The SizeConstraintSetId of the SizeConstraintSet that you want to get. SizeConstraintSetId is returned by CreateSizeConstraintSet and by ListSizeConstraintSets .

Destructuring the Response

getSizeConstraintSetResponse

Arguments

:: Int
gscsrsResponseStatus

-> GetSizeConstraintSetResponse

Creates a value of GetSizeConstraintSetResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gscsrsSizeConstraintSet - Information about the SizeConstraintSet that you specified in the GetSizeConstraintSet request. For more information, see the following topics: * SizeConstraintSet : Contains SizeConstraintSetId , SizeConstraints , and Name * SizeConstraints : Contains an array of SizeConstraint objects. Each SizeConstraint object contains FieldToMatch , TextTransformation , ComparisonOperator , and Size * FieldToMatch : Contains Data and Type

gscsrsResponseStatus - -- | The response status code.

data GetSizeConstraintSetResponse

See: getSizeConstraintSetResponse smart constructor.

Response Lenses

gscsrsSizeConstraintSet :: Lens' GetSizeConstraintSetResponse (Maybe SizeConstraintSet)

Information about the SizeConstraintSet that you specified in the GetSizeConstraintSet request. For more information, see the following topics: * SizeConstraintSet : Contains SizeConstraintSetId , SizeConstraints , and Name * SizeConstraints : Contains an array of SizeConstraint objects. Each SizeConstraint object contains FieldToMatch , TextTransformation , ComparisonOperator , and Size * FieldToMatch : Contains Data and Type

gscsrsResponseStatus :: Lens' GetSizeConstraintSetResponse Int

- | The response status code.