Network.AWS.WAFRegional.ListResourcesForWebACL

Description

Returns an array of resources associated with the specified web ACL.

Creating a Request

listResourcesForWebACL

Arguments

:: Text
lrfwaWebACLId

-> ListResourcesForWebACL

Creates a value of ListResourcesForWebACL with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lrfwaWebACLId - The unique identifier (ID) of the web ACL for which to list the associated resources.

data ListResourcesForWebACL

See: listResourcesForWebACL smart constructor.

Request Lenses

lrfwaWebACLId :: Lens' ListResourcesForWebACL Text

The unique identifier (ID) of the web ACL for which to list the associated resources.

Destructuring the Response

listResourcesForWebACLResponse

Arguments

:: Int
lrfwarsResponseStatus

-> ListResourcesForWebACLResponse

Creates a value of ListResourcesForWebACLResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lrfwarsResourceARNs - An array of ARNs (Amazon Resource Names) of the resources associated with the specified web ACL. An array with zero elements is returned if there are no resources associated with the web ACL.

lrfwarsResponseStatus - -- | The response status code.

data ListResourcesForWebACLResponse

See: listResourcesForWebACLResponse smart constructor.

Response Lenses

lrfwarsResourceARNs :: Lens' ListResourcesForWebACLResponse [Text]

An array of ARNs (Amazon Resource Names) of the resources associated with the specified web ACL. An array with zero elements is returned if there are no resources associated with the web ACL.

lrfwarsResponseStatus :: Lens' ListResourcesForWebACLResponse Int

- | The response status code.