Network.AWS.WAFRegional.UpdateXSSMatchSet

Description

Inserts or deletes XssMatchTuple objects (filters) in an XssMatchSet . For each XssMatchTuple object, you specify the following values:

Action : Whether to insert the object into or delete the object from the array. To change a XssMatchTuple , you delete the existing object and add a new one.

FieldToMatch : The part of web requests that you want AWS WAF to inspect and, if you want AWS WAF to inspect a header, the name of the header.

TextTransformation : Which text transformation, if any, to perform on the web request before inspecting the request for cross-site scripting attacks.

You use XssMatchSet objects to specify which CloudFront requests you want to allow, block, or count. For example, if you're receiving requests that contain cross-site scripting attacks in the request body and you want to block the requests, you can create an XssMatchSet with the applicable settings, and then configure AWS WAF to block the requests.

To create and configure an XssMatchSet , perform the following steps:

Submit a CreateXssMatchSet request.

Use GetChangeToken to get the change token that you provide in the ChangeToken parameter of an UpdateIPSet request.

Submit an UpdateXssMatchSet request to specify the parts of web requests that you want AWS WAF to inspect for cross-site scripting attacks.

For more information about how to use the AWS WAF API to allow or block HTTP requests, see the AWS WAF Developer Guide .

Creating a Request

updateXSSMatchSet

Arguments

:: Text
uxmsXSSMatchSetId

-> Text
uxmsChangeToken

-> UpdateXSSMatchSet

Creates a value of UpdateXSSMatchSet with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

uxmsXSSMatchSetId - The XssMatchSetId of the XssMatchSet that you want to update. XssMatchSetId is returned by CreateXssMatchSet and by ListXssMatchSets .

uxmsChangeToken - The value returned by the most recent call to GetChangeToken .

uxmsUpdates - An array of XssMatchSetUpdate objects that you want to insert into or delete from a XssMatchSet . For more information, see the applicable data types: * XssMatchSetUpdate : Contains Action and XssMatchTuple * XssMatchTuple : Contains FieldToMatch and TextTransformation * FieldToMatch : Contains Data and Type

data UpdateXSSMatchSet

A request to update an XssMatchSet .

See: updateXSSMatchSet smart constructor.

Request Lenses

uxmsXSSMatchSetId :: Lens' UpdateXSSMatchSet Text

The XssMatchSetId of the XssMatchSet that you want to update. XssMatchSetId is returned by CreateXssMatchSet and by ListXssMatchSets .

uxmsChangeToken :: Lens' UpdateXSSMatchSet Text

The value returned by the most recent call to GetChangeToken .

uxmsUpdates :: Lens' UpdateXSSMatchSet [XSSMatchSetUpdate]

An array of XssMatchSetUpdate objects that you want to insert into or delete from a XssMatchSet . For more information, see the applicable data types: * XssMatchSetUpdate : Contains Action and XssMatchTuple * XssMatchTuple : Contains FieldToMatch and TextTransformation * FieldToMatch : Contains Data and Type

Destructuring the Response

updateXSSMatchSetResponse

Arguments

:: Int
uxmsrsResponseStatus

-> UpdateXSSMatchSetResponse

Creates a value of UpdateXSSMatchSetResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

uxmsrsChangeToken - The ChangeToken that you used to submit the UpdateXssMatchSet request. You can also use this value to query the status of the request. For more information, see GetChangeTokenStatus .

uxmsrsResponseStatus - -- | The response status code.

data UpdateXSSMatchSetResponse

The response to an UpdateXssMatchSets request.

See: updateXSSMatchSetResponse smart constructor.

Response Lenses

uxmsrsChangeToken :: Lens' UpdateXSSMatchSetResponse (Maybe Text)

The ChangeToken that you used to submit the UpdateXssMatchSet request. You can also use this value to query the status of the request. For more information, see GetChangeTokenStatus .

uxmsrsResponseStatus :: Lens' UpdateXSSMatchSetResponse Int

- | The response status code.